Crate testnet_interlay_runtime_parachain

The testnet runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use btc_relay::bitcoin;

pub use constants::currency::*;

pub use constants::time::*;

pub use primitives;

pub use primitives::CurrencyId::ForeignAsset;

pub use primitives::CurrencyId::LendToken;

pub use primitives::CurrencyId::Token;

Modules

api

constants

A set of constant values used in the testnet runtime.

token_distribution

xcm_config

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AdjustmentVariable

The adjustment variable of the runtime. Higher values will cause TargetBlockFullness to change the fees more rapidly.

AssetAuthority

BalanceWrapper

a wrapper around a balance, used in RPC to workaround a bug where using u128 in runtime-apis fails. See https://github.com/paritytech/substrate/issues/4641

BaseCallFilter

BasicDeposit

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1 and adding 0.

BlockHashCount

BlockNumberToBalance

Burn

CollatorPotId

CollatorSelectionAccount

CurrencyHooks

DealWithFees

DealWithRewards

DustRemovalWhitelist

EmissionPeriod

EnactmentPeriod

EscrowAnnuityAccount

EscrowAnnuityPalletId

EscrowBlockRewardProvider

ExecutiveBody

ExistentialDeposits

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1 and adding 0.

FastTrackVotingPeriod

FeeAccount

FeePalletId

FieldDeposit

FixedU128

A fixed point number representation in the range. Fixed Point 128 bits unsigned, range = [0.000000000000000000, 340282366920938463463.374607431768211455]

GenesisConfig

GetDepositBase

GetDepositFactor

GetMaxSignatories

GetNativeCurrencyId

GetRelayChainCurrencyId

GetWrappedCurrencyId

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

InflationPeriod

IssueRequest

LaunchPeriod

LoansPalletId

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxCandidates

MaxExpectedValue

MaxInvulnerables

MaxLocks

MaxPeriod

MaxProposals

MaxRegistrars

MaxScheduledPerBlock

MaxSpend

MaxSubAccounts

MaxVestingSchedules

MaxVotes

MaximumMultiplier

MaximumSchedulerWeight

MinCandidates

MinVestedTransfer

MinimumDeposit

MinimumMultiplier

Minimum amount of the multiplier. This value cannot be too low. A test case should ensure that combined with AdjustmentVariable, we can recover from the minimum. See multiplier_can_grow_from_zero.

MinimumPeriod

Offset

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParachainBlocksPerBitcoinBlock

Perbill

A fixed point representation of a number in the range [0, 1].

Period

Permill

A fixed point representation of a number in the range [0, 1].

PreimageBaseDepositz

PreimageByteDeposit

PreimageByteDepositz

ProposalBond

ProposalBondMaximum

ProposalBondMinimum

RedeemRequest

ReplaceRequest

ReservedDmpWeight

ReservedXcmpWeight

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

Span

SpendPeriod

SubAccountDeposit

SupplyAccount

SupplyPalletId

TargetBlockFullness

The portion of the NORMAL_DISPATCH_RATIO that we adjust the fees with. Blocks filled less than this will decrease the weight and more will increase.

TechnicalCommitteeMaxMembers

TechnicalCommitteeMaxProposals

TechnicalCommitteeMotionDuration

TotalWrapped

TransactionByteFee

TreasuryAccount

TreasuryPalletId

UncleGenerations

VaultAnnuityAccount

VaultAnnuityPalletId

VaultBlockRewardProvider

VaultRegistryAccount

VaultRegistryPalletId

Version

VotingPeriod

Weight

Enums

BtcRelayCall

Contains one variant per dispatchable that can be called by an extrinsic.

DispatchClass

A generalized group of dispatch types.

Everything

A [Contains] implementation that contains every value.

IssueEvent

The event emitted by this pallet.

NominationEvent

The event emitted by this pallet.

Nothing

A [Contains] implementation that contains no value.

OriginCaller

RedeemEvent

The event emitted by this pallet.

ReplaceEvent

The event emitted by this pallet.

RuntimeCall

RuntimeEvent

StatusCode

Enum indicating the status of the BTC Parachain.

TimestampCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 12 second average block time.

TARGET_SPACING

Target Spacing: 10 minutes (600 seconds)

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

CurrencyInfo

Get

A trait for querying a single value from a type.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

get_all_module_accounts

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssetRegistry

AssetRegistryConfig

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

BTCRelay

BTCRelayConfig

Balance

Balance of an account.

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

ClientsInfo

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

We allow root and the Relay Chain council to execute privileged collator selection operations.

CumulusXcm

Currency

Democracy

DemocracyConfig

DmpQueue

Escrow

EscrowAnnuity

EscrowRewards

Executive

Executive: handles dispatch to the various modules.

Fee

FeeConfig

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Identity

Issue

IssueConfig

Liquidity

Loans

LoansConfig

LockIdentifier

An identifier for a lock. Used for disambiguating different locks so that they can be individually replaced or removed.

Moment

An instant or duration in time.

Multisig

Nomination

NominationConfig

Nonce

Index of a transaction in the chain. 32-bit should be plenty.

Oracle

OracleConfig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Preimage

Rate

Ratio

Redeem

RedeemConfig

Replace

ReplaceConfig

Scheduler

Security

SecurityConfig

Session

SessionConfig

Shortfall

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedFixedPoint

The signed fixed point type.

SignedInner

The Inner type of the SignedFixedPoint.

SlowAdjustingFeeUpdate

Sudo

SudoConfig

Supply

SupplyConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tokens

TokensConfig

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UnknownTokens

UnsignedFixedPoint

The unsigned fixed point type.

UnsignedInner

The Inner type of the UnsignedFixedPoint.

Utility

VaultAnnuity

VaultRegistry

VaultRegistryConfig

VaultRewards

VaultStaking

Vesting

VestingConfig

XTokens

XcmpQueue